Refactor vec1 implementation
