--- glm/gtc/vec1.inl
@@ -1,2 +1,0 @@
-/// @ref gtc_vec1
-/// @file glm/gtc/vec1.inl
